+++ src/main/java/org/example/mollyapi/user/dto/SignUpReqDto.java
@@ -6,8 +6,8 @@ import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
-import org.example.mollyapi.user.entity.Auth;
-import org.example.mollyapi.user.entity.Password;
+import org.example.mollyapi.user.auth.entity.Auth;
+import org.example.mollyapi.user.auth.entity.Password;
 import org.example.mollyapi.user.entity.User;
 import org.example.mollyapi.user.type.Role;
 import org.example.mollyapi.user.type.Sex;
